requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
click==6.7                # via python-dotenv
django-anymail[mailgun]==1.0
django-registration-redux==1.8
django==1.11.6
idna==2.6                 # via requests
python-dotenv==0.7.1
pytz==2017.2              # via django
requests==2.18.4          # via django-anymail
six==1.11.0               # via django-anymail
urllib3==1.22             # via requests
